requirements.txt
-i https://pypi.org/simple
bcrypt==4.1.3; python_version >= '3.7'
cffi==1.16.0; platform_python_implementation != 'PyPy'
colorama==0.4.6; sys_platform == 'win32'
cryptography==42.0.8; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
logzero==1.7.0
paramiko==3.4.0; python_version >= '3.6'
pycparser==2.22; python_version >= '3.8'
pycryptodome==3.20.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pymongo==4.8.0; python_version >= '3.8'
pynacl==1.5.0; python_version >= '3.6'
pysftp==0.2.9
python-dotenv==1.0.1; python_version >= '3.8'
